requirements.txt
Django==1.8.4
django-filter==0.11.0
django-pagination-py3==1.1.1
pytz==2018.7
